+++ vue-app/src/components/FormNavigation.vue
@@ -9,13 +9,18 @@
       >
         Previous
       </button>
+      <wallet-widget
+        class="float-right"
+        v-if="!currentUser && currentStep === finalStep"
+        :isActionButton="true"
+      />
       <button
-        v-if="currentStep < finalStep"
-        @click="handleNext"
+        v-else
+        @click="handleStepNav(currentStep + 1, true)"
         class="btn-primary float-right"
         :disabled="!isStepValid"
       >
-        {{ currentStep === finalStep - 1 ? 'Finish' : 'Next' }}
+        {{ currentStep === finalStep ? 'Confirm' : 'Next' }}
       </button>
     </div>
   </div>
@@ -25,32 +30,25 @@
 import Vue from 'vue'
 import Component from 'vue-class-component'
 import { Prop } from 'vue-property-decorator'
+import { User } from '@/api/user'
+import WalletWidget from '@/components/WalletWidget.vue'
 
-@Component
+@Component({
+  components: {
+    WalletWidget,
+  },
+})
 export default class FormNavigation extends Vue {
   @Prop() currentStep!: number
   @Prop() finalStep!: number
   @Prop() steps!: string[]
   @Prop() isStepValid!: boolean
   @Prop() callback!: (updateFurthest?: boolean) => void
-  @Prop() handleStepNav!: () => void
+  @Prop() handleStepNav!: (step: number, updateFurthest?: boolean) => void
   @Prop() isNavDisabled!: boolean
 
-  // TODO is this needed? Why do we pass `callback` & `handleStepNav`?
-  handleNext(): void {
-    // Save form data (first saves when user hits Next after first step)
-    this.callback(true) // "true" checks to update furthest step
-    // Navigate forward
-    this.$router.push({
-      name: 'join-step',
-      params: {
-        step: this.steps[this.currentStep + 1],
-      },
-    })
-  }
-
-  handlePrev(): void {
-    this.callback()
+  get currentUser(): User | null {
+    return this.$store.state.currentUser
   }
 }
 </script>
